+++ app/channels/application_cable/connection.rb
@@ -5,15 +5,7 @@ module ApplicationCable
     identified_by :current_user
 
     def connect
-      set_current_user || reject_unauthorized_connection
-    end
-
-    private
-
-    def set_current_user
-      return unless (session = Session.find_by(id: cookies.signed[:session_id]))
-
-      self.current_user = session.user
+      self.current_user = env['warden']&.user || reject_unauthorized_connection
     end
   end
 end