+++ app/controllers/pages_controller.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 class PagesController < ApplicationController
-  allow_unauthenticated_access
+  skip_before_action :authenticate_user!
 
   def home
-    redirect_to dashboard_path if authenticated?
+    redirect_to dashboard_path if user_signed_in?
   end
 end