+++ app/controllers/registrations_controller.rb
@@ -1,26 +1,13 @@
 # frozen_string_literal: true
 
 class RegistrationsController < ApplicationController
-  allow_unauthenticated_access
+  skip_before_action :authenticate_user!
 
   def new
-    @user = User.new
+    redirect_to new_user_registration_path
   end
 
   def create
-    @user = User.new(user_params)
-
-    if @user.save
-      start_new_session_for(@user)
-      redirect_to dashboard_path, notice: t('auth.registration.welcome')
-    else
-      render :new, status: :unprocessable_content
-    end
-  end
-
-  private
-
-  def user_params
-    params.expect(user: %i[email_address username password password_confirmation])
+    redirect_to new_user_registration_path
   end
 end