+++ app/controllers/tournament/rounds_controller.rb
@@ -2,7 +2,7 @@
 
 module Tournament
   class RoundsController < ApplicationController
-    before_action :authenticate!
+    before_action :authenticate_user!
     before_action :set_tournament
 
     def index
@@ -16,9 +16,7 @@ module Tournament
 
     private
 
-    def authenticate!
-      redirect_to new_session_path unless Current.user
-    end
+    # Devise provides authentication; Current.user is set at ApplicationController
 
     def set_tournament
       @tournament = ::Tournament::Tournament.find(params[:tournament_id])