+++ test/system/games_test.rb
@@ -71,9 +71,9 @@ class GamesTest < ApplicationSystemTestCase
   private
 
   def login_as(user)
-    visit new_session_path
-    fill_in 'email_address', with: user.email_address
-    fill_in 'password', with: 'password'
+    visit new_user_session_path
+    fill_in 'Email', with: user.email
+    fill_in 'Password', with: 'password'
     click_on I18n.t('auth.login')
   end
 end